LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.arcTo.nonfinite-expected.txt

2d.path.arcTo.nonfinite
arcTo() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS arcTo() with Infinity/NaN is ignored
